refact(Application/HeadlessProvider): move models to the provider :recycle::
