+++ src/Core/Application/Contracts/Providers/HeadlessProvider/DTOs/Actions/Click.cs
@@ -1,0 +1,8 @@
+namespace AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs.Actions;
+
+using AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs.Actions.Shared;
+
+public class Click : BaseAction
+{
+
+}
