+++ src/Core/Application/Contracts/Providers/HeadlessProvider/DTOs/Actions/Open.cs
@@ -1,0 +1,7 @@
+namespace AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs.Actions;
+
+public class Open
+{
+  public bool headless { get; set; }
+
+}
