+++ src/Core/Application/Contracts/Providers/HeadlessProvider/DTOs/ElementChecker.cs
@@ -1,0 +1,6 @@
+namespace AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs;
+
+public class ElementChecker
+{
+  public string? text { get; set; }
+}
