+++ src/Core/Application/Contracts/Providers/HeadlessProvider/DTOs/Selectors/GetElement.cs
@@ -1,0 +1,8 @@
+namespace AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs.Selectors;
+
+using AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs.Selectors.Shared;
+
+public class GetElement : BaseSelector
+{
+
+}
