+++ src/Core/Application/Contracts/Providers/HeadlessProvider/DTOs/Selectors/GetText.cs
@@ -1,0 +1,7 @@
+namespace AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs.Selectors;
+
+using AnyTime.Core.Application.Contracts.Providers.HeadlessProvider.DTOs.Selectors.Shared;
+
+public class GetText : BaseSelector
+{
+}
