--- src/Core/Application/Models/Headless/ActionCheck.cs
@@ -1,7 +1,0 @@
-namespace AnyTime.Core.Application.Models.Headless;
-
-public class ActionCheck
-{
-  public string? text { get; set; }
-
-}
