--- src/Core/Application/Models/Headless/Open.cs
@@ -1,7 +1,0 @@
-namespace AnyTime.Core.Application.Models.Headless;
-
-public class Open
-{
-  public bool headless { get; set; }
-
-}
