--- src/Core/Application/Models/Headless/Selectors/Click.cs
@@ -1,5 +1,0 @@
-namespace AnyTime.Core.Application.Models.Headless.Selectors;
-public class Click : Selector
-{
-
-}
