--- src/Core/Application/Models/Headless/Selectors/FillInput.cs
@@ -1,5 +1,0 @@
-namespace AnyTime.Core.Application.Models.Headless.Selectors;
-public class FillInput : Selector
-{
-  public string content { get; set; }
-}
