--- src/Core/Application/Models/Headless/Selectors/GetAttribute.cs
@@ -1,5 +1,0 @@
-namespace AnyTime.Core.Application.Models.Headless.Selectors;
-public class GetAttribute : Selector
-{
-  public string attributeName { get; set; }
-}
