--- src/Core/Application/Models/Headless/Selectors/GetElement.cs
@@ -1,5 +1,0 @@
-namespace AnyTime.Core.Application.Models.Headless.Selectors;
-public class GetElement : Selector
-{
-
-}
