--- src/Core/Application/Models/Headless/Selectors/GetText.cs
@@ -1,5 +1,0 @@
-namespace AnyTime.Core.Application.Models.Headless.Selectors;
-
-public class GetText : Selector
-{
-}
